Vss2Git

+++ Арканоид/RealTime.h
@@ -3,14 +3,14 @@
 
 class RealTime {
 private:
-	static double _freq, _time_offset;
+	static float _freq, _time_offset;
 	static bool _initialized;
 
 	static void _init();
 public:
-	static double getTime();
-	static void setTime(double time);
-	static void wait(double delay_time);
+	static float getTime();
+	static void setTime(float time);
+	static void wait(float delay_time);
 };
 
 #endif	// REALTIME_H_INCLUDED
